MAR-1745 added spacer slicer | renamed Boilerplate Grid to Card Grid
