+++ .slicemachine/assets/client/prismicSlices/pageParts/SpacerSlice/index.stories.js
@@ -1,0 +1,26 @@
+import MyComponent from '../../../../../../client/prismicSlices/pageParts/SpacerSlice';
+import SliceZone from 'vue-slicezone'
+
+export default {
+  title: 'client/prismicSlices/pageParts/SpacerSlice'
+}
+
+
+export const _DefaultSlice = () => ({
+  components: {
+    MyComponent,
+    SliceZone
+  },
+  methods: {
+    resolve() {
+      return MyComponent
+    }
+  },
+  data() {
+    return {
+      mock: {"variation":"default-slice","name":"Default slice","slice_type":"spacer_slice","items":[],"primary":{"height":"50px","color":"#555555"},"id":"_DefaultSlice"}
+    }
+  },
+  template: '<SliceZone :slices="[mock]" :resolver="resolve" />'
+})
+_DefaultSlice.storyName = 'Default slice'
