+++ .slicemachine/assets/client/prismicSlices/pageParts/SpacerSlice/mocks.json
@@ -1,0 +1,12 @@
+[
+  {
+    "variation": "default-slice",
+    "name": "Default slice",
+    "slice_type": "spacer_slice",
+    "items": [],
+    "primary": {
+      "height": "50px",
+      "color": "#555555"
+    }
+  }
+]
