+++ client/prismicSlices/pageParts/index.js
@@ -1,10 +1,11 @@
 // This file is generated by Prismic local builder. Update with care!
 
-export { default as BoilerplatesGridSlice } from './BoilerplatesGridSlice'
+export { default as CardGridSlice } from './CardGridSlice'
 export { default as CreateProjectsSlice } from './CreateProjectsSlice'
 export { default as CtaBannerSlice } from './CtaBannerSlice'
 export { default as CustomerLogoSlice } from './CustomerLogoSlice'
-export { default as ProjectsGridSlice } from './ProjectsGridSlice'
 export { default as KeyMetricsSlice } from './KeyMetricsSlice'
+export { default as ProjectsGridSlice } from './ProjectsGridSlice'
+export { default as SpacerSlice } from './SpacerSlice'
 export { default as StartScreen } from './StartScreen'
 export { default as TextSlice } from './TextSlice'
